[shardformer] rename policy file name
